pyproject.toml
[project]
name = "gslidegen"
version = "0.1.0"
description = "Tableau workbook export and PDF processing utilities"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "pdf2image>=1.17.0",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "requests",
]

[tool.setuptools.packages.find]
where = ["src"]
